requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
antspyx==0.5.3
async-timeout==4.0.3
attrs==24.2.0
bitsandbytes==0.44.1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
docker-pycreds==0.4.0
docstring_parser==0.16
filelock==3.16.1
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2024.9.0
gitdb==4.0.11
GitPython==3.1.43
hydra-core==1.3.2
idna==3.10
importlib_resources==6.4.5
Jinja2==3.1.5
joblib==1.4.2
jsonargparse==4.33.2
kiwisolver==1.4.7
lightning==2.4.0
lightning-utilities==0.11.8
markdown-it-py==3.0.0
MarkupSafe==3.0.1
matplotlib==3.9.2
mdurl==0.1.2
monai==1.3.2
mpmath==1.3.0
multidict==6.1.0
networkx==3.4.1
nibabel==5.3.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.77
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
packaging==24.1
pandas==2.2.3
patsy==0.5.6
pillow==11.0.0
platformdirs==4.3.6
propcache==0.2.0
protobuf==5.28.2
psutil==6.0.0
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytorch-lightning==2.4.0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
rich==13.9.2
scikit-learn==1.5.2
scipy==1.14.1
sentry-sdk==2.16.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
statsmodels==0.14.4
sympy==1.13.3
tensorboardX==2.6.2.2
threadpoolctl==3.5.0
torch==2.4.1
torchmetrics==1.4.3
torchvision==0.19.1
tqdm==4.66.5
triton==3.0.0
typeshed_client==2.7.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wandb==0.18.3
webcolors==24.8.0
yarl==1.15.3
tensorflow==2.16.1
tensorflow-probability[tf]==0.24.0
monai-generative
lpips
diffusers
transformers
pyarrow
gdown
antspyx
antspynet
